Scripts/StockApp-v2/backend/requirements.txt
django>=4.2
djangorestframework>=3.15
django-cors-headers>=4.3
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
statsmodels>=0.14.0
pmdarima>=2.0.0
scipy>=1.11.0
plotly>=5.17.0
